prospecting_engine/requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
tavily-python==0.3.1
reportlab==4.4.3
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
